Create todo.java
This To-Do List Application in Java allows users to create and manage a list of tasks. The program provides a simple command-line interface where the user can interact with the application by selecting options to add, view, or remove tasks.

Key Features:
Add Task: The user can input a task, which will be added to the list of tasks.

View Tasks: Displays all the tasks in the list, numbered for easy reference.

Remove Task: The user can remove a specific task by entering its number.

Exit: The program terminates when the user selects the exit option.